fix: harden core engine — btree balancing, ARC cache eviction, LIMIT in fallback paths, BOCPD run-length tracking, NaN guards, and mutex poison resilience
Multiple correctness and robustness fixes across the core engine crates:

btree: Remove early-return in balance_nonroot when all_cells is empty so
that balance_shallower can still collapse empty interior roots back into
leaves. Widen split_overflowing_root child_count range from 3..=8 to
2..=8 to handle small-cell root splits. Cap freelist trunk leaf count to
prevent buffer overflows on small pages.

core/connection: Fix LIMIT clause being silently dropped in GROUP BY,
JOIN, MATCH, and subquery fallback SELECT paths — the bound struct's
limit was consumed by bind_placeholders but never re-applied after row
materialization. Fix indentation of execute_with_materialized_views.

pager/arc_cache: Replace evict_lru_from_t1/t2 with delete_lru_from_t1/t2
that remove pages entirely instead of moving them to ghost lists, fixing
a capacity-overflow loop when T1 alone fills capacity. Update ARC cache
tests to match corrected eviction semantics.

mvcc/bocpd: Track explicit run_length in RunEntry instead of using
vec index as proxy — after pruning low-probability entries the index
no longer corresponds to actual run length, causing stale MAP estimates
and false changepoint detections.

parser/semantic: Add USING-column disambiguation so that columns joined
via USING are no longer flagged as ambiguous. Support non-reserved
keywords after '.' in qualified column references. Fix function arity:
likelihood(X,P) is 2-arg, likely/unlikely are 1-arg, count is exactly 1.
Properly propagate INSERT alias into upsert DO UPDATE scope.

planner/stats: Add NaN guards and zero-row-count checks in selectivity
interpolation to prevent NaN propagation through query cost estimates.

vdbe: Prefix STRICT type-check error with SQLITE_CONSTRAINT_DATATYPE
marker for parity with C SQLite error reporting. Replace manual range
check with idiomatic contains(). Add unit test for the marker.

vdbe/vectorized_agg: Eliminate redundant double-update of accumulators
by storing a global ordered index in the group_map bucket instead of
maintaining a parallel group_order vector.

vfs: Replace all .expect("...lock poisoned") with
.unwrap_or_else(|e| e.into_inner()) across shm, unix inode table,
and io_uring tests to prevent cascading panics when a prior thread
poisoned a mutex.

wal: Expose frame_offset as pub(crate) and use it in group_commit
instead of duplicating the offset arithmetic. Simplify checksum test
formatting.

datetime: Use saturating arithmetic throughout Julian Day conversions
and timediff to prevent overflow on extreme date values.

+++ crates/fsqlite-mvcc/src/physical_merge.rs
@@ -674,7 +674,7 @@ pub fn apply_patch(
     }
 
     // Return cells sorted by rowid (table) or key bytes (index) for canonical order
-    let usable = base.page_size.usable(base.reserved_per_page) as u32;
+    let usable = base.page_size.usable(base.reserved_per_page);
     let mut cells: Vec<ParsedCell> = cell_map.into_values().collect();
     cells.sort_by(|a, b| {
         if let (Some(ra), Some(rb)) = (a.rowid, b.rowid) {
